synthetic_image_generation_pipeline/requirements.txt
numpy
matplotlib
Pillow
seaborn
pandas
plotly
bokeh
# Optional image export backends
kaleido
selenium
